fix bug in GetBulk behavior when last key is missing
GetBulk uses GETQ to get all keys except the last key
which is retrieved using GET.  The expected behavior
is to simply return no information about keys that
are missing.  Previously, if the last key in the list
was missing, we mistakenly returned an error, instead
of simply ignoring it, as we do for all the other keys.

This change checks the status of response if there was an error,
and if the status was KEY_ENOENT that error is not returned.

Change-Id: Iaf1a38a24721cd57081c0c382dd2840bf8bce14c