MB-13575 Rest request to goxdcr timed out 1. for delete remote cluster reference and replication, the handler code (doDeleteRemoteClusterRequest, DeleteReplication) do a query for the remote cluster reference\replication on metakv, then call metakv to delete. The reason to query is to get the detail informations for audit event. ReplicationSepcSvc.DelReplicationSpec method and RemoteClusterService.DelRemoteCluster would do another query on metakv to get the revision before delete. so it is a three round trip to metakv, which is quite heavy.
The fix is to let ReplicationSpecSvc.DelReplicationSpec and RemoteClusterService.DelRemoteCluster to return the metadata deleted to save one round trip.

MB-13630 GOXDCR: Replication resumed should be started right way
In the code, the resumed replication is treated the same as the failed-and-retried replication to let it wait for retry_interval timed out before starting it. It should start it right way.

Change-Id: I6571b3439708ba31ca807272eba4b6d7e23902b5
Reviewed-on: http://review.couchbase.org/47358
Reviewed-by: Yu Sui <ysui68@gmail.com>
Tested-by: Xiaomei Zhang <xiaomei@couchbase.com>