MB-27415 n1ql connection leak under load
The problem is that go-couchbase/conn_pool.go:Return() will close a gomemcached connection if this is deemed to be unhealthy.
A connection is deemed to be unhealthy if the result from the last request was considered to be fatal for the connection, and IsFatal() returned false for very few errors, which probably was correct at the beginning of time.
Snag is, loads of new errors have been added to mc_constants.go, and very few of these are fatal, but IsFatal() has never been amended.
Added a map[Status] bool to show what errors are fatals and what are not.

Change-Id: Ibc6d1420ec8274a2b8fc9d5f087541485b5504d7