MB-35542: Only use scramSha if remote cluster is compatible.
We set httpAuthMech based on what GetSecuritySettingAndDefaultPoolInfo() returns.
The logic in this routine will return:
- Plain if not half encryption encryption and no certificate:
- ScramSha if we half encryption and we can connect with ScramSha
- https if
  * not half encryption or half but ScramSha connect failed, and
  * we have a certificate and https connect worked, and
  * cluster is not ScramSha compatible

Note that when cluster is not ScramSha compatible and half encrypt, we will return
https or ScramSha based on the current host. This causes httpAuthMech change in
RemoteClusterRef and pipeline retart.

We need to add the logic to use https if cluster is not ScramSha compatible.

Change-Id: Ie2ba41e60bc9321747dc3e1a3df2cd3ee77d3d46
Reviewed-on: http://review.couchbase.org/114198
Reviewed-by: Neil Huang <neil.huang@couchbase.com>
Reviewed-by: John Liang <johliang00@yahoo.com>
Tested-by: Lilei Chen <lilei.chen@couchbase.com>