Issue: MB-37134 XDCR stuck in Pending.
RCA:
The issue is located around the in cache updated. The ReplicationSpecVal cache can be updated in local UI operation and in the metaKV update (through Cluster). The in memory object  (cache) has two parts 1) configuration part called as spec and operational part called as replicationStatus. These two parts can be updated concurrently by two different threads. The cache update uses the Read Modify Write(RMW) mechanism where threads can be preempted after the read and mean time other thread can go and change the other part of the object and subsequent write of RMW can overwrite with old stuff. The following are the line of events would have gotten the system into this problematic state

 1) The node got a metaKV update for a replication delete and the serializer routine (thread) of the corresponding topic would have signaled the updater routine
 2) The updater routine would have started executing StopPipeline() while it was resetting or cleaning up the ReplicationStatus object
 3) The node would have got a create request either through local UI or through the cluster which would have tried to update the cache with new spec by cacheSpec()
 4) The cacheSpec() routine which is using RMW to update the cache would have preempted or raced after the read
 5) The updater routine would have exited after continuing the rest of the StopPipeline() which would have cleared the DerivedObj (ReplicationSpecVal)
 6) The UI routine would have continued the write which would have overwritten the stale DerivedObj (ReplicationSpecVal).

Fix:
The most of the cache updates are protected/serialized by the
ReplicaitonSpecService cache lock however the operational object(replicationStatus) update was not protected by the ReplicaitonSpecService cache lock which caused some inconsistency in the large cluster setup. This code change close the gap and all replication spec cache update are protected by single lock.

Added unit test to exercise the concurrent access of the replication object in the cache. Though the CBSE-7648 can’t be simulated with standalone unit test code, it is tested by implanting the temporary probes in the production code. The probes are not part of the changeset however this code will improve the code coverage.

Tests:
  1) Simulated the issue and tested the fix.
  2) Carried out regression testing by executing typical replication create,delete and update operations in multi node cluster.
  3) Added some documents and confirmed the document tranfer.

Change-Id: I94578a302feb15a481e92c0663e915aa1e965860
Reviewed-on: http://review.couchbase.org/118835
Reviewed-by: John Liang <johliang00@yahoo.com>
Reviewed-by: Neil Huang <neil.huang@couchbase.com>
Reviewed-by: Lilei Chen <lilei.chen@couchbase.com>
Reviewed-by: <senthilkumar.narayanasamy@couchbase.com>
Tested-by: <senthilkumar.narayanasamy@couchbase.com>