MB-37819 - Remove/Suppress XDCR warnings for non-Data Service nodes
Analysis: It is a regression in 6.5.0. The one of the changes got in part of MB-35119 regressed this. The code has a logic to filter the error ErrorNoSourceNozzle during the UI reporting. The filter compares the error event with raw error message of  ErrorNoSourceNozzle but the event message was modified with prefix (topic...)

Fix: filter is changed to do substring comparison.

Change-Id: I14e245af9782a31330eab5865375f184f033ee50
Testing: Manually tested with 3 node cluster.
Reviewed-on: http://review.couchbase.org/123323
Reviewed-by: Neil Huang <neil.huang@couchbase.com>
Reviewed-by: Lilei Chen <lilei.chen@couchbase.com>
Reviewed-by: <senthilkumar.narayanasamy@couchbase.com>
Tested-by: <senthilkumar.narayanasamy@couchbase.com>