MB-38378 - BP 6.5.1 XDCR changes growing - "stream already exists for this vbucket"
patch-3: Refactored the comments in the code.

patch-2: resets the vm-stream state with previous state upon stream request
send failure.

RCA: It is a timing issue, could happen in a heavily loaded system with
large number of buckets. The replication instantiation process tries to
create a DCP stream for every vBucket by sending the create request and
moves the state of the vBucket to INIT state. The response for the create
request will set the state as ACTIVE and start processing the data.
In certain condition thread which sets the INIT state can be preempted and
executed after the response process. Though stream was created, the preempted
thread can overwrite the ACTIVE state with INIT and lock the particular vBucket from streaming

Fix: Set the INIT state prior to sending the stream create request, it
will protect from the racing in response path.

Testing: The issue was reproduced with two buckets by loading the
documents and confirmed the difference of the document count. Tested the fix with
10 buckets.

Change-Id: Id9c86b256b3e5404c8a3a29db0d8b0f1086acff1
Reviewed-on: http://review.couchbase.org/123600
Reviewed-by: <senthilkumar.narayanasamy@couchbase.com>
Tested-by: <senthilkumar.narayanasamy@couchbase.com>
(cherry picked from commit 8c40585085d1b9939ff64d4cd776d932c63e3480)
Reviewed-on: http://review.couchbase.org/124071
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Lilei Chen <lilei.chen@couchbase.com>