MB-41239  - User Intent heuristic is incorrect for full-encryption when ref does not provide port number
If users do not enter a specific admin port in the hostname field when creating a remote cluster reference,
XDCR tags it with 8091 automatically. This has always been the case even if for Full Encryption,
just in case that if somehow the user sets 8091 to be the ssl rest port.

With user intent (MB-37626), XDCR is supposed to check the target cluster's node list and see if the alternate
address matches the user's entered hostname.
This means that if user did not enter :18091, but chooses Full-Encryption, and the alternate address section
of the remote cluster has :18091 as the SSL port, the user intent will be incorrect.

Change-Id: If907cd95726f9f3818257c237e0c6e053083e6f0
Reviewed-on: http://review.couchbase.org/c/goxdcr/+/136976
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Neil Huang <neil.huang@couchbase.com>
Reviewed-by: Neil Huang <neil.huang@couchbase.com>