MB-41201 - User Intent heuristic is incorrect for full-encryption when ref does not provide port number
If users do not enter a specific admin port in the hostname field when creating a remote cluster reference,
XDCR tags it with 8091 automatically. This has always been the case even if for Full Encryption,
just in case that if somehow the user sets 8091 to be the ssl rest port.

With user intent (MB-37626), XDCR is supposed to check the target cluster's node list and see if the alternate
address matches the user's entered hostname.
This means that if user did not enter :18091, but chooses Full-Encryption, and the alternate address section
of the remote cluster has :18091 as the SSL port, the user intent will be incorrect.

Change-Id: I600bcded3afb1bd9ad05109609f946c068c6dea9
Reviewed-on: http://review.couchbase.org/c/goxdcr/+/135217
Tested-by: Neil Huang <neil.huang@couchbase.com>
Reviewed-by: Neil Huang <neil.huang@couchbase.com>