MB-42188: Fix set back logic
If source cluster is slow, it may perform a merge later than target.
Target may have already merged the same documents and did more
operations, resulting in a document with smaller CAS but dominates
source document. The code assume that the target is a merged document
with no additional change, which is incorrect.

Change-Id: Id376fee1693b49ce40eef7bc9d6d10d0e948079f
Reviewed-on: http://review.couchbase.org/c/goxdcr/+/138583
Reviewed-by: Neil Huang <neil.huang@couchbase.com>
Tested-by: Lilei Chen <lilei.chen@couchbase.com>