MB-44683 - tgt manifests can be lost when pipeline restarts due to target errors
It is possible that when a pipeline resumes, all the VB's are able to
resume a DCP stream. This leads to XDCR declaring the pipeline "ready
for checkpoint" as part of checkpoint manager's isCheckpointAllowed()

However, if XMEM is stuck or things are timing out, it is possible for
certain VBs in ThroughSeqnoTracker to not have any data flow. Then, the
next time ckpt mgr performs a checkpoint, it may retrieve 0 for the
target manifest ID and commit that into a checkpoint. If this happens
too often it'll render all the ckpts with 0 as their target manifest IDs
This will then lead to collections Manifest service throwing away all
manifests for the target.

Change-Id: Id1eb999150d9a41e608df85e4a968de32f568a08
Reviewed-on: http://review.couchbase.org/c/goxdcr/+/147659
Tested-by: Neil Huang <neil.huang@couchbase.com>
Reviewed-by: Neil Huang <neil.huang@couchbase.com>
Reviewed-by: Lilei Chen <lilei.chen@couchbase.com>