MB-44597 - Concurrent R/W operations on VBTasksMap
Even though backfillRequestHandler's job is to act as a singleton to
handle incoming requests so that there is no concurrency issues, it does
not have the jurisdiction outside of that. For example, other components
can request the backfill replication spec, which is provided by the
backfill replication service. The specs need to share the same VBTaskMap
instance with backfillReqHandler so that the VBTaskmap can be updated
when requests come in. This ultimately causes concurrency issues.

The fix is simple in concept: modify the VBTasksMap structure to be
wrapped around with a mutex. However, since the structure embeds other
types of structures that also need wrapping, the solution extends to
those other structures as well. The callers need to ensure to use the
right type of RW mutex depending on its needed operation

Change-Id: Ie6106d4c8bd0bc8814ea4064100389346d907656
Reviewed-on: http://review.couchbase.org/c/goxdcr/+/148256
Tested-by: Neil Huang <neil.huang@couchbase.com>
Reviewed-by: Neil Huang <neil.huang@couchbase.com>