MB-43352 - Part 5: Cleaning up dismissed events and hook up with manifests monitoring
This changeset hooks up the events manager with backfill manager, which
is responsible to figure out what mappings are being backfilled or
cleaned up whenever manifests changes occur.

A few clean ups take place here:

1)
Whenever a broken mapping is backfilled, we take this information and
try to match it against what the user has dismissed, and remove those
dismissed entries. The thinking is that if a brokenmapping is repaired,
then the users' dismissed actions are considered out of date since
they were technically dismissing earlier versions of the broken mappings

2)
Whenever source scopes or collections are removed, backfillMgr will
find these out and notify the corresponding pipeline's eventsMgr.
Each eventMgr will ensure that outdated entries are removed from
its own cached brokenMap and dismissedMap

3)
When a pipeline is stopped, we will also "forget" all of the users'
dismissed actions

In addition to above, a few bug fixes took place:
  - Change data types to pointers whenever possible to avoid potentially
    create duplicate references when passing things around
  - Hook up pipeline manager actions into serializer to ensure that mgt
    operations per pipeline do not block and are serialized in nature
  - Added some more test cases and expand live rediscover test case 2 to
    validate dismiss operation and also the two scenarios above on
    how to "reset" previous dismissed actions

Change-Id: Ica3a85b7993c3f3395e156be8bfdfec4dd9cbf45
Reviewed-on: http://review.couchbase.org/c/goxdcr/+/149396
Tested-by: Neil Huang <neil.huang@couchbase.com>
Reviewed-by: Neil Huang <neil.huang@couchbase.com>
Reviewed-by: Lilei Chen <lilei.chen@couchbase.com>