MB-43352 - Part 6: error handling for slow internal processing
It is possible that traversing the data structures when brokenMap is big
could be expensive and slow. Callers will expect that XDCR does not
hang even while it is churning away

This changeset proposes an "eventual consistent" model for brokenMap
and dismiss actions. If an operation is taking too long, then XDCR
will return back to the caller before the operations are finished

Change-Id: I9acb3e1501c991b1bc69f02b41f67afbc71f9aa7
Reviewed-on: http://review.couchbase.org/c/goxdcr/+/149428
Tested-by: Neil Huang <neil.huang@couchbase.com>
Reviewed-by: Neil Huang <neil.huang@couchbase.com>
Reviewed-by: Lilei Chen <lilei.chen@couchbase.com>