MB-45983 - Part 4: Hook up bucket topology service to backfill manager and handler
Backfill Manager and handler used to retrieve VB maps and lists with
immediate feedback. But moving to bucket topology service means
that VBs list may be delayed. When things are delayed, then there's
a potential race when certain VBs are added or removed.

There are some code path changes to support "catching up"
if VBs changes were delayed

In the future, when backfill specs are shared between nodes,
then this needs to be revisited

Change-Id: I34930db1fbc1daa1ccc165dd56eac86d8ceb0bf6
Reviewed-on: http://review.couchbase.org/c/goxdcr/+/156374
Reviewed-by: Lilei Chen <lilei.chen@couchbase.com>
Reviewed-by: Neil Huang <neil.huang@couchbase.com>
Tested-by: Neil Huang <neil.huang@couchbase.com>