MB-47570 - concurrent settings map RW issue
This exporting flag into multiple values function call will issue a write on
the settingsMap object itself. The issue is exacerbated by the fact that
bucket topology service has multiple instances of watchers firing concurrently
where all of them will access settingsmap

The exporting of flags only matter to outward facing calls, which should not be
often. So the fix is to clone it before modifying it

Change-Id: I5c810f3b29619053c9d0fc70ecf1b006e7cc63d3
Reviewed-on: http://review.couchbase.org/c/goxdcr/+/158238
Tested-by: Neil Huang <neil.huang@couchbase.com>
Reviewed-by: Neil Huang <neil.huang@couchbase.com>