MB-48280: Documents not getting replicated to remote
In a previous change, MyMemcachedAddr is changed to local loopback
address if cluster encryption is strict, which sets MyKVNodes to loopback
address. However, serverVBMap key was not changed. When the loopback
address is used as key in GetSourceVBMap(), no source VB can be found.

This change fix the inconsistency and always uses loopback as current node
key in serverVBMap if encryption is strict. This ensures that
serverVBMap can be searched correctly and the serverAddr can be used to
connect to source.

Also ensure 8901 is not used when connecting to remote and only TLS is
used when cluster uses strict encryption.

Change-Id: Ic87ddf3d2e4aeea80f260451894dc3bd0093230d
Reviewed-on: http://review.couchbase.org/c/goxdcr/+/160389
Reviewed-by: Neil Huang <neil.huang@couchbase.com>
Reviewed-by: Lilei Chen <lilei.chen@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Lilei Chen <lilei.chen@couchbase.com>
Reviewed-on: http://review.couchbase.org/c/goxdcr/+/160650