MB-48323 - return latest cached value for KV check or memcached address
ns_server's pools/nodes endpoint is already considered a expensive call
that we should avoid calling. MB-44823 already added a caching layer,
but if the last call failed, the error is also saved and fed back to
callers who request the pools/nodes endpoint information

For calls of MyKVNodes() and MyMemcachedAddr() that are used repeatedly
but have little chance of changing, it could return its last cached
information, instead of returning the last cached error in the error
case. Since if the last cached is used, the logs will spam with a bunch
of errors giving the illusion that a lot of the ns_server calls are
failing when actually the one call failed and the cached error is
being used repeatedly downstream

Change-Id: Ie6ee7d857d97f1de502569dc19d1a14408e18797
Reviewed-on: http://review.couchbase.org/c/goxdcr/+/160988
Reviewed-by: Lilei Chen <lilei.chen@couchbase.com>
Reviewed-by: Neil Huang <neil.huang@couchbase.com>
Tested-by: Neil Huang <neil.huang@couchbase.com>