MB-47672 - return latest cached value for KV check or memcached address
ns_server's pools/nodes endpoint is already considered a expensive call
that we should avoid calling. MB-44823 already added a caching layer,
but if the last call failed, the error is also saved and fed back to
callers who request the pools/nodes endpoint information

For calls of MyKVNodes() and MyMemcachedAddr() that are used repeatedly
but have little chance of changing, it could return its last cached
information, instead of returning the last cached error in the error
case. Since if the last cached is used, the logs will spam with a bunch
of errors giving the illusion that a lot of the ns_server calls are
failing when actually the one call failed and the cached error is
being used repeatedly downstream

Change-Id: Ie0c84d177f5a6f0cd25c884ea0e8a8fa4be643dd
Reviewed-on: http://review.couchbase.org/c/goxdcr/+/164679
Well-Formed: Restriction Checker
Reviewed-by: Lilei Chen <lilei.chen@couchbase.com>
Reviewed-by: Neil Huang <neil.huang@couchbase.com>
Tested-by: Neil Huang <neil.huang@couchbase.com>