MB-49432: Ensure in-build library paths are first on link path
If we link against the install directory (which is where third-party
libraries are located) first, then when doing incremental builds during
development, we'll try to link against libraries from earlier builds
such as libjseval. If those libraries happen to have a breaking API
change, or introduce new APIs that goxdcr depends on, the build will
break with a strange link error.

Change-Id: I9e4b9568a9fd227414ce934c61b4efce0b250309
Reviewed-on: http://review.couchbase.org/c/goxdcr/+/165315
Reviewed-by: Lilei Chen <lilei.chen@couchbase.com>
Reviewed-by: Neil Huang <neil.huang@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>