MB-49972: test framework fix
In a later change, ns_server's pools/default returns a cluster
group of name "Group 0".
This throws off the shell script as the space in the JSON will
cause it to not parse correctly by default.

A fix is needed to ensure the script failing a node out can work

Change-Id: I252fa43ede4c9f4494328a41f64f10387374a34f
Reviewed-on: https://review.couchbase.org/c/goxdcr/+/167187
Tested-by: Neil Huang <neil.huang@couchbase.com>
Reviewed-by: Neil Huang <neil.huang@couchbase.com>