MB-49434: Checkpoint cache optimization cache to prevent repeatedly hitting simple_store
In a busy system, repeatedly hitting simple_store seems to adversely
affect performance. This changeset introduces a in-memory caching
layer for ckpts in between XDCR and simple_store

When ckpts are first loaded, they are instantiated in the cache.
When pipeline performs ckpt in each VB, the cache is updated.
When ckptSvc is asked to provide ckpt data, it will look at the cache first.
If cache is missed, then it'll look at simple_store

Change-Id: Ic90724560a581783eb729155ea2a048aac026c19
Reviewed-on: https://review.couchbase.org/c/goxdcr/+/165869
Reviewed-by: Lilei Chen <lilei.chen@couchbase.com>
Reviewed-by: Neil Huang <neil.huang@couchbase.com>
Tested-by: Neil Huang <neil.huang@couchbase.com>