MB-50416: prevent duplicate push merge from locking up
From the logs, we see that backfillSpec.VBTaskMap can be modified
and so there is a possibility that .Contains() check is not fully
dependable.

It is possible that a task within a VBTaskList is being remerged
If so, then it may lock with itself.

Backfill Manager should also be aware of the differences in between
push and pull requests from different peer nodes, so that
it doesn't forget what push requests from VB master that it has
already handled. Currently, if a push request was handled
and a pull request comes in, the pull request overwrites the push
request and so the duplicate push is forgotten, leading to the
whole Contains() miss and merge-with-itself issue

Change-Id: If6ba8242ad4f13cacd03080ffb22752ad1a2f555
Reviewed-on: https://review.couchbase.org/c/goxdcr/+/169537
Reviewed-by: Lilei Chen <lilei.chen@couchbase.com>
Reviewed-by: Neil Huang <neil.huang@couchbase.com>
Tested-by: Neil Huang <neil.huang@couchbase.com>