MB-50854: do not skip checkpointing when stopping backfill pipeline with callback
At least for the majority of the cases, stopping backfill pipeline
should cause a checkpoint to take place.
Before this fix, the skipCkpt option is defaulted to true, which
can cause interruption to long running backfill tasks to not make
progress.

Certain scenarios, however, may be ok to skip checkpoint for
performance reasons, such as handling rollback to 0

Added test case for checking for checkpoints and validation
framework

Change-Id: Ib710b1ed161d487bb8d76152f702ad663db35e2e
Reviewed-on: https://review.couchbase.org/c/goxdcr/+/170609
Reviewed-by: Lilei Chen <lilei.chen@couchbase.com>
Reviewed-by: Neil Huang <neil.huang@couchbase.com>
Tested-by: Neil Huang <neil.huang@couchbase.com>