MB-54317: Increase js evaluator quota
With the V8 getting upgraded to v10.7.21
1.52 MB is the heap size for an empty isolate.

Keeping the heap quota to as low as 1.5 MB triggers muliple
unnecessary GC runs every 30 seconds. Increase the heap quota
to avoid such runs.

Increase to 15MB for now until more is tested to be safer
Note the increase should not come into affect if JSWorker is
turned off

Change-Id: Iee28c5b4208f0dd8488d380704a7b86b559cafc2
Reviewed-on: https://review.couchbase.org/c/goxdcr/+/182116
Reviewed-by: Lilei Chen <lilei.chen@couchbase.com>
Reviewed-by: Neil Huang <neil.huang@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Neil Huang <neil.huang@couchbase.com>