MB-55412: backfillMgr to initialized source manifest
Prior to p2p (7.1), a newly added rebalanced-in node
always starts from Seqno 0. But, with P2P, it can now
pull checkpoints from other nodes. If backfillMgr initiates
and does not "fast-forward" to this point in time, then it
can miss out on diffing the right manifests and thus
miss out on backfills

A few test cases were modified to capture some potential dummy
backfills that could get started because of this change

Change-Id: Ia47494d686fc74bc8c70219839ac485ed8ddd68c
Reviewed-on: https://review.couchbase.org/c/goxdcr/+/186450
Tested-by: Neil Huang <neil.huang@couchbase.com>
Reviewed-by: Neil Huang <neil.huang@couchbase.com>
Reviewed-by: Lilei Chen <lilei.chen@couchbase.com>