MB-57382: backfillMgr to initialized source manifest
Prior to p2p (7.1), a newly added rebalanced-in node
always starts from Seqno 0. But, with P2P, it can now
pull checkpoints from other nodes. If backfillMgr initiates
and does not "fast-forward" to this point in time, then it
can miss out on diffing the right manifests and thus
miss out on backfills

A few test cases were modified to capture some potential dummy
backfills that could get started because of this change

Change-Id: I427be3da25b723da138ad564743fd9a10cdced20
Reviewed-on: https://review.couchbase.org/c/goxdcr/+/196966
Tested-by: Neil Huang <neil.huang@couchbase.com>
Reviewed-by: Neil Huang <neil.huang@couchbase.com>
Well-Formed: Restriction Checker