MB-57180: Fix issues causing changes_left not going to 0
There are two problems that causes OsoSession.isDoneNoLock() to always return false and changes_left to not
go down to zero

First, during OSO, seqno is not in order. vbClonedTracker still treat it as sorted when searching, causing it
to not find the cloned event. So each DataSent is process as handled, causing s.seqnosHandledCnt to be greater
than s.seqnosFromDCPCnt and OsoSession.isDoneNoLock() to return false.

Second, the DataSent event may come before DataCloned. It will not find the cloned event. So each DataSent is
process as handled, causing s.seqnosHandledCnt to be greater than s.seqnosFromDCPCnt and OsoSession.isDoneNoLock()
to return false.

This change ensures that the events are processed in the right order, and the list is searched correctly.

Change-Id: Ia3e18d08f4dcefcfe8e6c156de5a825e7bf6e3e4
Reviewed-on: https://review.couchbase.org/c/goxdcr/+/191954
Tested-by: Lilei Chen <lilei.chen@couchbase.com>
Reviewed-by: Neil Huang <neil.huang@couchbase.com>