MB-63378: statsMgr to return trad vs global stats
Before being able to create global checkpoints,
statsMgr needs to be able to handle a generic way to
populate stats to be ready for checkpointing.
The checkpoint logic stays the same, and so does the way
checkpoint manager invokes statsMgr to populate VBspecific
stats.

The difference now is how statsMgr internally handle the
population of stats to be fed back to checkpoint manager

Change-Id: I15ba47e9a489c3b304a98ad002356551ef7b48b5