MB-63700: globalCkpt - fix checkpoint sorting algorithm
This changeset also adds two new test cases:
1. A 128-1024 broken map live rediscover test case
2. A peer to peer checkpoint sorting + broken map test case

This changeset found and fixed an issue in throughSeqnoTrackerSvc
because system events were still working off of the target's VB
which caused a panic.

There are more work to be done in the next changeset that
will address broken mapping issues properly to ensure that
p2p tests will pass. For now, the p2p tests are added but are
not passing

Change-Id: I08239aa6587debc29b453dd50554b817d5a05771
Reviewed-on: https://review.couchbase.org/c/goxdcr/+/216873
Reviewed-by: Darshan N A <darshan.na@couchbase.com>
Tested-by: Neil Huang <neil.huang@couchbase.com>
Reviewed-by: Neil Huang <neil.huang@couchbase.com>