MB-63467: p2p pull with broken map issues found and fixed
First, test cases were added (1e) to test and ensure
that broken maps are supposed to be transferred with variable
vb mode

As part of the debugging, various data were captured and used
as part of unit test throughout the code path to ensure that
p2p code works properly. A few things may have been fixed in
that path.

Finally, the issue was found when broken map "cleanup" code
was taking place, due to the incorrect population of sha
by using the traditional sha instead of the global shas.
That led to mistakenly removing brokenmap post transfer
on the side of the node that pulled

Change-Id: I16a2d1a456a847a618ec5fef034115be04fbdf75