MB-65120: Switch back cLogTracker truncate function to use binary search
- TL;DR:
In the original implementation of truncate for cLogTracker, the implementation used linear search for truncate which mainly was based on the fact that the list would be not sorted to use binary search. However going deep into the golang sort.Search function, it turns out we can use binary search eventhough the lists are not sorted.

- Explantion:
The golang sort.Search(n, f) function is documented as:

Search uses binary search to find and return the smallest index i in [0, n) at which f(i) is true, assuming that on the range [0, n), f(i) == true implies f(i+1) == true. That is, Search requires that f is false for some (possibly empty) prefix of the input range [0, n) and then true for the (possibly empty) remainder; Search returns the first true index. If there is no such index, Search returns n. (Note that the "not found" return value is not -1 as in, for instance, strings.Index.) Search calls f(i) only for i in the range [0, n).

The important bit being:

Search requires that f is false for some (possibly empty) prefix of the input range [0, n) and then true for the (possibly empty) remainder

For our use-case, the seqno_lists will be such that it is sorted till through_seqno and unsorted after that. Eg: seqno_list=[1, 2, 3, 5, 4], through_seqno=3. Our f(i) is defined as seqno_list[i] >= through_seqno. Therefore we abide by the Sort function’s prerequisite that we have the seqno_list split into 2 parts indexes [0, i] and [i+1, n), with i-th index element being through_seqno and seqno_list[j] < through_seqno for all j in the range [0, i] AND seqno_list[k] >= through_seqno for all k in the range [i+1, n).

Therefore, eventhough in our use-case, the lists are not sorted, we honour the invariant defined for the binary search

Change-Id: Ifa15960416c780f985565db11964c3b398d16230
Reviewed-on: https://review.couchbase.org/c/goxdcr/+/226688
Tested-by: Sumukh Bhat <sumukh.bhat@couchbase.com>
Reviewed-by: Neil Huang <neil.huang@couchbase.com>